refactor(web): break the import cycles noImportCycles found
Biome resolves imports more thoroughly than madge did, so replacing
`quality:circular` surfaced ten cycles the old gate never reported. All of
them came from two knots.

`HeaderManager` imported every concrete header and also exported the
`SidebarHeaderButton` those headers render, so each header imported its own
parent back — directly for Calendar/Chat, and through `NotificationCenter`
for Todos. The button is a shared primitive rather than part of the
orchestrator, so it moves to its own module and the five consumers point at
it instead.

`todoStore` called `startWorkflowPolling` on creation while the polling
module wrote results back through the store. The store now emits a signal
that the global workflow listener subscribes to, so the dependency runs one
way. Polling is registered from the listener's effect rather than at import
time, which keeps the wiring visible: it is armed exactly while the listener
that cancels it on a WebSocket delivery is mounted. `startWorkflowPolling`
is no longer exported, since the signal is the only entry point.

With the cycles gone the rule moves from `warn` to `error`, so
`quality:circular` is an enforced gate again rather than a report.
