+++ apps/web/src/components/layout/headers/ChatHeader.tsx
@@ -3,8 +3,8 @@
 import { Kbd } from "@heroui/kbd";
 import { BubbleChatAddIcon, PinIcon, SearchIcon } from "@icons";
 import Link from "next/link";
-import { SidebarHeaderButton } from "@/components/layout/headers/HeaderManager";
 import ModelSelectorDevControls from "@/components/layout/headers/ModelSelectorDevControls";
+import { SidebarHeaderButton } from "@/components/layout/headers/SidebarHeaderButton";
 import { prepareNewChat } from "@/features/chat/utils/newChatNavigation";
 import { NotificationCenter } from "@/features/notification/components/NotificationCenter";
 import { usePlatform } from "@/hooks/ui/usePlatform";
