+++ apps/web/src/stores/todoStore.ts
@@ -1,7 +1,7 @@
 import { createTodoStore } from "@shared/todos";
 
 import { todoApi } from "@/features/todo/api/todoApi";
-import { startWorkflowPolling } from "@/features/todo/hooks/useTodoWorkflowGlobalListener";
+import { emitTodoCreated } from "@/features/todo/utils/todoCreatedSignal";
 import { toast } from "@/lib/toast";
 
 export const useTodoStore = createTodoStore(todoApi, {
@@ -10,6 +10,6 @@ export const useTodoStore = createTodoStore(todoApi, {
     error: (msg) => toast.error(msg),
     info: (msg) => toast.info(msg),
   },
-  onTodoCreated: (todoId) => startWorkflowPolling(todoId),
+  onTodoCreated: (todoId) => emitTodoCreated(todoId),
   devtoolsName: "todo-store",
 });
