+++ biome.json
@@ -153,7 +153,7 @@
         "noGlobalEval": "error"
       },
       "nursery": {
-        "noImportCycles": "warn"
+        "noImportCycles": "error"
       }
     }
   },
